+++ src/static.ts
@@ -1,3 +1,4 @@
 export const DEFAULT_NAMESPACE = "__def__"
+export const ERROR_TYPE_RESPONSE = "______error_type_response____"
 
 export const IS_CONTENT_SCRIPT = !chrome.tabs