+++ src/types.ts
@@ -1,6 +1,6 @@
 export type ReceiverReply<T = any> = (msg: T) => void
 export type ReceiverHandle<T = any, Q = any> = (data: T, reply: ReceiverReply<Q>) => void
-export type MessageHandle<T = any, Q = any> = (msg: T, reply: (response: Q) => void, tab: Required<chrome.tabs.Tab>, sender: chrome.runtime.MessageSender, port: chrome.runtime.Port) => void
+export type MessageHandle<T = any, Q = any> = (msg: T, tab: Required<chrome.tabs.Tab>, sender: chrome.runtime.MessageSender, port: chrome.runtime.Port) => Q|Promise<Q>|void
 
 export type MessageHandleTemplate = Record<string, MessageHandle>
 export type UnionMessageHandleTemplate = Record<string, MessageHandleTemplate>
@@ -22,3 +22,8 @@ export type PassingData<T = any> = {
 }
 
 export type MiddlewareData<T = any> = PassingData<T>
+
+export type ErrorResponse = {
+  type: string;
+  error: any;
+}